+++ src/main/java/de/dm/intellij/liferay/language/osgi/ComponentInspectionToolProvider.java
@@ -1,0 +1,14 @@
+package de.dm.intellij.liferay.language.osgi;
+
+import com.intellij.codeInspection.InspectionToolProvider;
+import org.jetbrains.annotations.NotNull;
+
+public class ComponentInspectionToolProvider implements InspectionToolProvider {
+    @NotNull
+    @Override
+    public Class[] getInspectionClasses() {
+        return new Class[] {
+                ComponentServiceInheritanceInspection.class
+        };
+    }
+}