+++ testdata/de/dm/intellij/liferay/language/osgi/ComponentServiceInheritanceInspectionTest/MyComponent_fixed.java
@@ -1,0 +1,8 @@
+import org.osgi.service.component.annotations.Component;
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
+
+@Component(
+    service = MVCActionCommand.class
+)
+public class MyComponent implements MVCActionCommand {
+}