+++ testdata/de/dm/intellij/liferay/language/osgi/ComponentServiceInheritanceInspectionTest/ValidSelfComponent.java
@@ -1,0 +1,7 @@
+import org.osgi.service.component.annotations.Component;
+
+@Component(
+    service = ValidSelfComponent.class
+)
+public class ValidSelfComponent {
+}